change create session test to use deferred promises
